Update ChaturbateRecorder.py
Redid the fix to the original file. 
Added some small debugging in that it will write the output from the chaturbate model list. 
Removed the looping through the model list and replaced it with code that will iterate through the model list each time it gets the list, this is done to prevent cloudflare from locking you out.

Added a file (record.txt) to just force recording of a model without checking the online model list. (so when all else fails...)